# schemas/api.yaml
openapi: 3.0.3
info:
  title: Workers HMD API
  description: API for managing Health Mediated Deployments for Cloudflare Workers
  version: 1.0.0
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: "https://workers-hmd.{cloudflare-workers-subdomain}.workers.dev/api"
    variables:
      cloudflare-workers-subdomain:
        default: replace-me
security:
  - BasicAuth: []
paths:
  /plan:
    get:
      summary: Get deployment plan
      description: Retrieve the current deployment plan with stages and SLOs
      operationId: getPlan
      tags:
        - Plan
      responses:
        200:
          description: Plan retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
        404:
          description: Plan not found
    post:
      summary: Update deployment plan
      description: Create or update the deployment plan configuration
      operationId: updatePlan
      tags:
        - Plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Plan"
      responses:
        200:
          description: Plan updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
        400:
          description: Plan is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /release:
    get:
      summary: List releases
      description: Get a list of all releases including current and completed ones
      operationId: listReleases
      tags:
        - Release
      parameters:
        - name: limit
          in: query
          description: Maximum number of releases to return (default 50, max 100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          description: Number of releases to skip for pagination (default 0)
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: since
          in: query
          description: Filter releases created after this timestamp (ISO 8601 format)
          required: false
          schema:
            type: string
            format: date-time
            example: '2024-01-01T00:00:00Z'
        - name: until
          in: query
          description: Filter releases created before this timestamp (ISO 8601 format)
          required: false
          schema:
            type: string
            format: date-time
            example: '2024-12-31T23:59:59Z'
        - name: state
          in: query
          description: Filter releases by state
          required: false
          schema:
            type: string
            enum: [not_started, running, done_successful, done_stopped_manually, done_failed_slo]
      responses:
        200:
          description: List of releases retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Release"
        400:
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create release
      description: Create a new release. Only one release can be staged at a time
      operationId: createRelease
      tags:
        - Release
      responses:
        200:
          description: Release created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Release"
        409:
          description: A release is already staged
        400:
          description: Release cannot be created from the plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /release/active:
    get:
      summary: Get active release
      description: Retrieve the active release, or null if no active release exists
      operationId: getActiveRelease
      tags:
        - Release
      responses:
        200:
          description: Success (returns active release or null)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Release"
                  - type: "null"
                description: Active release object if one exists, null otherwise
    post:
      summary: Control active release execution
      description: Start or stop the active release
      operationId: controlActiveRelease
      tags:
        - Release
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - command
              properties:
                command:
                  type: string
                  enum:
                    - start
                    - stop
                  description: The command to execute on the active release
                account_id:
                  type: string
                  description: Cloudflare Account ID
                api_token:
                  type: string
                  description: Cloudflare API Token for authentication
      responses:
        200:
          description: start or stop was successful
          content:
            text/plain:
              schema:
                type: string
                example: "Release started successfully"
        400:
          description: The provided command is not valid for the release in its current state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: No active release found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete active release
      description: Delete the active release if it is in 'not_started' state
      operationId: deleteActiveRelease
      tags:
        - Release
      responses:
        200:
          description: Release deleted
        400:
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: No active release found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        409:
          description: Release has to be in a "not_started" state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /release/{releaseId}:
    get:
      summary: Get release details
      description: Retrieve release state, plan record, and stage information
      operationId: getRelease
      tags:
        - Release
      parameters:
        - $ref: "#/components/parameters/releaseId"
      responses:
        200:
          description: Release details retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Release"
        404:
          description: Release not found
  /stage/{stageId}:
    get:
      summary: Get stage status
      description: Get the current state of a specific stage
      operationId: getStage
      tags:
        - Stage
      parameters:
        - $ref: "#/components/parameters/stageId"
      responses:
        200:
          description: Returns the Stage info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseStage"
        404:
          description: Stage not found
    post:
      summary: Progress stage
      description: Advance a stage by approving or denying it
      operationId: progressStage
      tags:
        - Stage
      parameters:
        - $ref: "#/components/parameters/stageId"
      requestBody:
        required: true
        content:
          application/text:
            schema:
              type: string
              enum:
                - approve
                - deny
      responses:
        200:
          description: Successfully progressed the stage
        400:
          description: Invalid command
        404:
          description: Stage not found
  /worker/versions:
    post:
      summary: Get worker versions
      description: Proxy request to fetch worker versions from Cloudflare API
      operationId: getWorkerVersions
      tags:
        - Worker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - worker_name
                - account_id
                - api_token
              properties:
                worker_name:
                  type: string
                  description: Name of the Cloudflare Worker
                account_id:
                  type: string
                  description: Cloudflare Account ID
                api_token:
                  type: string
                  description: Cloudflare API Token for authentication
      responses:
        200:
          description: Worker versions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  result:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorkerVersion"
        400:
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Invalid API token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Worker not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /worker/deployments:
    post:
      summary: Get worker deployments
      description: Proxy request to fetch worker deployments from Cloudflare API
      operationId: getWorkerDeployments
      tags:
        - Worker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - worker_name
                - account_id
                - api_token
              properties:
                worker_name:
                  type: string
                  description: Name of the Cloudflare Worker
                account_id:
                  type: string
                  description: Cloudflare Account ID
                api_token:
                  type: string
                  description: Cloudflare API Token for authentication
      responses:
        200:
          description: Worker deployments retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  result:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorkerDeployment"
        400:
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Invalid API token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Worker not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
  parameters:
    stageId:
      name: stageId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{8}$'
        example: '0fcdec2eb49045cc8b481c15e0f1277c'
    releaseId:
      name: releaseId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{8}$'
        example: '0fcdec2eb49045cc8b481c15e0f1277c'
  schemas:
...
    Release:
      type: object
      readOnly: true
      example:
        id: "b2a2c3d4e5f6a1b2d3d5e5f6a1b2c3d5"
        state: "not_started"
        plan_record:
          $ref: "#/components/schemas/Plan"
        old_version: "fa2c3d4e5a6a1b2d3d5e5f6a1b2c3d5"
        new_version: "e1a2c3d4e5f6a1b2d3d5e5f6a1b2c3d5"
        stages:
          type: array
          items:
            $ref: "#/components/schemas/StageRef"
        time_started:
          type: string
          format: date-time
        time_elapsed:
          type: integer
        time_done:
          type: string
          format: date-time
      required:
        - id
        - state
        - plan_record
        - stages
        - old_version
        - new_version
        - time_created
        - time_started
        - time_elapsed
        - time_done
      properties:
        id:
          type: string
        state:
          type: string
          enum:
            - not_started
            - running
            - done_stopped_manually
            - done_failed_slo
            - done_successful
            - error
        plan_record:
          $ref: "#/components/schemas/Plan"
        old_version:
          type: string
        new_version:
          type: string
        stages:
          type: array
          items:
            $ref: "#/components/schemas/StageRef"
        time_created:
          type: string
          format: date-time
        time_started:
          type: string
          format: date-time
        time_elapsed:
          type: integer
        time_done:
          type: string
          format: date-time
...
    Plan:
      type: object
      example:
        stages:
          - $ref: "#/components/schemas/PlanStage"
        slos:
          - $ref: "#/components/schemas/SLO"
        worker_name: "my-worker"
      required:
        - stages
        - slos
        - worker_name
        - polling_fraction
      properties:
        stages:
          type: array
          items:
            $ref: "#/components/schemas/PlanStage"
        slos:
          type: array
          items:
            $ref: "#/components/schemas/SLO"
        polling_fraction:
          type: number
          minimum: 0.1
          maximum: 1.0
          description: Fraction of soak time to use for SLO polling rate (0.1 = 1/10, 1.0 = full time)
          example: 0.5
        time_last_saved:
          type: string
          format: date-time
          readOnly: true
        worker_name:
          type: string
...
    Error:
      type: object
      description: Standard error response format
      example:
        message: "Invalid request"
        ok: false
      required:
        - message
        - ok
      properties:
        message:
          type: string
          description: Error message describing what went wrong
        ok:
          type: boolean
          description: Always false for error responses
          enum: [false]